+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon -L ",
+    "start": "node ./index.js ",
     "test": "mocha -w ./tests/**/*.spec.js"
   },
   "author": "Henry",
