+++ src/routes/index.js
@@ -6,12 +6,11 @@ const getAllUsers = require("./usersRouter");
 const getAllServices = require("./ServicesRouter");
 const getAllPartners = require("./partnersRouter");
 const getAllBills = require("./billsRouter");
-const paymentValidation = require("./paymentRouter");
 const getAllCities = require("./createIdRouter");
 const reservationRouter = require("./reservationRouter");
 const routeSearchHotels = require("./SearchHotels");
 
-const cors = require('cors');
+const cors = require("cors");
 const router = Router();
 
 // Configurar los routers
@@ -24,7 +23,6 @@ router.use("/users", getAllUsers);
 router.use("/services", getAllServices);
 router.use("/partners", getAllPartners);
 router.use("/bills", getAllBills);
-router.use("/payment", paymentValidation);
 router.use("/cities", getAllCities);
 router.use("/reservations", reservationRouter);
 router.use("/search", routeSearchHotels);
